+++ modules/AWSPowerShell/Cmdlets/AmazonMediaConvert/Basic/Add-EMCResourceTag-Cmdlet.cs
@@ -22,44 +22,55 @@ using System.Management.Automation;
 using System.Text;
 using Amazon.PowerShell.Common;
 using Amazon.Runtime;
-using Amazon.ServerlessApplicationRepository;
-using Amazon.ServerlessApplicationRepository.Model;
+using Amazon.MediaConvert;
+using Amazon.MediaConvert.Model;
 
-namespace Amazon.PowerShell.Cmdlets.SAR
+namespace Amazon.PowerShell.Cmdlets.EMC
 {
     /// <summary>
-    /// Sets the permission policy for an application. See <a href="https://docs.aws.amazon.com/serverlessrepo/latest/devguide/access-control-resource-based.html#application-permissions">Application
-    /// Permissions</a> for the list of supported actions that can be used with this operation.
+    /// Tag a MediaConvert queue, preset, or job template. For information about these resource
+    /// types, see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html
     /// </summary>
-    [Cmdlet("Set", "SARApplicationPolicy", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
-    [OutputType("Amazon.ServerlessApplicationRepository.Model.ApplicationPolicyStatement")]
-    [AWSCmdlet("Calls the AWS Serverless Application Repository PutApplicationPolicy API operation.", Operation = new[] {"PutApplicationPolicy"})]
-    [AWSCmdletOutput("Amazon.ServerlessApplicationRepository.Model.ApplicationPolicyStatement",
-        "This cmdlet returns a collection of ApplicationPolicyStatement objects.",
-        "The service call response (type Amazon.ServerlessApplicationRepository.Model.PutApplicationPolicyResponse) can also be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
+    [Cmdlet("Add", "EMCResourceTag", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
+    [OutputType("None","System.String")]
+    [AWSCmdlet("Calls the AWS Elemental MediaConvert TagResource API operation.", Operation = new[] {"TagResource"})]
+    [AWSCmdletOutput("None or System.String",
+        "When you use the PassThru parameter, this cmdlet outputs the value supplied to the Arn parameter. Otherwise, this cmdlet does not return any output. " +
+        "The service response (type Amazon.MediaConvert.Model.TagResourceResponse) can be referenced from properties attached to the cmdlet entry in the $AWSHistory stack."
     )]
-    public partial class SetSARApplicationPolicyCmdlet : AmazonServerlessApplicationRepositoryClientCmdlet, IExecutor
+    public partial class AddEMCResourceTagCmdlet : AmazonMediaConvertClientCmdlet, IExecutor
     {
         
-        #region Parameter ApplicationId
+        #region Parameter Arn
         /// <summary>
         /// <para>
-        /// <para>The Amazon Resource Name (ARN) of the application.</para>
+        /// The Amazon Resource Name (ARN) of the resource that
+        /// you want to tag. To get the ARN, send a GET request with the resource name.
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]
-        public System.String ApplicationId { get; set; }
+        public System.String Arn { get; set; }
         #endregion
         
-        #region Parameter Statement
+        #region Parameter Tag
         /// <summary>
         /// <para>
-        /// <para>An array of policy statements applied to the application.</para>
+        /// The tags that you want to add to the resource. You
+        /// can tag resources with a key-value pair or with only a key.
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]
-        [Alias("Statements")]
-        public Amazon.ServerlessApplicationRepository.Model.ApplicationPolicyStatement[] Statement { get; set; }
+        [Alias("Tags")]
+        public System.Collections.Hashtable Tag { get; set; }
+        #endregion
+        
+        #region Parameter PassThru
+        /// <summary>
+        /// Returns the value passed to the Arn parameter.
+        /// By default, this cmdlet does not generate any output.
+        /// </summary>
+        [System.Management.Automation.Parameter]
+        public SwitchParameter PassThru { get; set; }
         #endregion
         
         #region Parameter Force
@@ -76,8 +87,8 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         {
             base.ProcessRecord();
             
-            var resourceIdentifiersText = FormatParameterValuesForConfirmationMsg("ApplicationId", MyInvocation.BoundParameters);
-            if (!ConfirmShouldProceed(this.Force.IsPresent, resourceIdentifiersText, "Set-SARApplicationPolicy (PutApplicationPolicy)"))
+            var resourceIdentifiersText = FormatParameterValuesForConfirmationMsg("Arn", MyInvocation.BoundParameters);
+            if (!ConfirmShouldProceed(this.Force.IsPresent, resourceIdentifiersText, "Add-EMCResourceTag (TagResource)"))
             {
                 return;
             }
@@ -91,10 +102,14 @@ namespace Amazon.PowerShell.Cmdlets.SAR
             // allow for manipulation of parameters prior to loading into context
             PreExecutionContextLoad(context);
             
-            context.ApplicationId = this.ApplicationId;
-            if (this.Statement != null)
+            context.Arn = this.Arn;
+            if (this.Tag != null)
             {
-                context.Statements = new List<Amazon.ServerlessApplicationRepository.Model.ApplicationPolicyStatement>(this.Statement);
+                context.Tags = new Dictionary<System.String, System.String>(StringComparer.Ordinal);
+                foreach (var hashKey in this.Tag.Keys)
+                {
+                    context.Tags.Add((String)hashKey, (String)(this.Tag[hashKey]));
+                }
             }
             
             // allow further manipulation of loaded context prior to processing
@@ -110,15 +125,15 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         {
             var cmdletContext = context as CmdletContext;
             // create request
-            var request = new Amazon.ServerlessApplicationRepository.Model.PutApplicationPolicyRequest();
+            var request = new Amazon.MediaConvert.Model.TagResourceRequest();
             
-            if (cmdletContext.ApplicationId != null)
+            if (cmdletContext.Arn != null)
             {
-                request.ApplicationId = cmdletContext.ApplicationId;
+                request.Arn = cmdletContext.Arn;
             }
-            if (cmdletContext.Statements != null)
+            if (cmdletContext.Tags != null)
             {
-                request.Statements = cmdletContext.Statements;
+                request.Tags = cmdletContext.Tags;
             }
             
             CmdletOutput output;
@@ -129,7 +144,9 @@ namespace Amazon.PowerShell.Cmdlets.SAR
             {
                 var response = CallAWSServiceOperation(client, request);
                 Dictionary<string, object> notes = null;
-                object pipelineOutput = response.Statements;
+                object pipelineOutput = null;
+                if (this.PassThru.IsPresent)
+                    pipelineOutput = this.Arn;
                 output = new CmdletOutput
                 {
                     PipelineOutput = pipelineOutput,
@@ -154,16 +171,16 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         
         #region AWS Service Operation Call
         
-        private Amazon.ServerlessApplicationRepository.Model.PutApplicationPolicyResponse CallAWSServiceOperation(IAmazonServerlessApplicationRepository client, Amazon.ServerlessApplicationRepository.Model.PutApplicationPolicyRequest request)
+        private Amazon.MediaConvert.Model.TagResourceResponse CallAWSServiceOperation(IAmazonMediaConvert client, Amazon.MediaConvert.Model.TagResourceRequest request)
         {
-            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "AWS Serverless Application Repository", "PutApplicationPolicy");
+            Utils.Common.WriteVerboseEndpointMessage(this, client.Config, "AWS Elemental MediaConvert", "TagResource");
             try
             {
                 #if DESKTOP
-                return client.PutApplicationPolicy(request);
+                return client.TagResource(request);
                 #elif CORECLR
                 // todo: handle AggregateException and extract true service exception for rethrow
-                var task = client.PutApplicationPolicyAsync(request);
+                var task = client.TagResourceAsync(request);
                 return task.Result;
                 #else
                         #error "Unknown build edition"
@@ -184,8 +201,8 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         
         internal partial class CmdletContext : ExecutorContext
         {
-            public System.String ApplicationId { get; set; }
-            public List<Amazon.ServerlessApplicationRepository.Model.ApplicationPolicyStatement> Statements { get; set; }
+            public System.String Arn { get; set; }
+            public Dictionary<System.String, System.String> Tags { get; set; }
         }
         
     }