+++ modules/AWSPowerShell/Cmdlets/AmazonMediaConvert/Basic/New-EMCQueue-Cmdlet.cs
@@ -62,6 +62,18 @@ namespace Amazon.PowerShell.Cmdlets.EMC
         public System.String Name { get; set; }
         #endregion
         
+        #region Parameter Tag
+        /// <summary>
+        /// <para>
+        /// The tags that you want to add to the resource. You
+        /// can tag resources with a key-value pair or with only a key.
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter]
+        [Alias("Tags")]
+        public System.Collections.Hashtable Tag { get; set; }
+        #endregion
+        
         #region Parameter Force
         /// <summary>
         /// This parameter overrides confirmation prompts to force 
@@ -93,6 +105,14 @@ namespace Amazon.PowerShell.Cmdlets.EMC
             
             context.Description = this.Description;
             context.Name = this.Name;
+            if (this.Tag != null)
+            {
+                context.Tags = new Dictionary<System.String, System.String>(StringComparer.Ordinal);
+                foreach (var hashKey in this.Tag.Keys)
+                {
+                    context.Tags.Add((String)hashKey, (String)(this.Tag[hashKey]));
+                }
+            }
             
             // allow further manipulation of loaded context prior to processing
             PostExecutionContextLoad(context);
@@ -116,6 +136,10 @@ namespace Amazon.PowerShell.Cmdlets.EMC
             if (cmdletContext.Name != null)
             {
                 request.Name = cmdletContext.Name;
+            }
+            if (cmdletContext.Tags != null)
+            {
+                request.Tags = cmdletContext.Tags;
             }
             
             CmdletOutput output;
@@ -183,6 +207,7 @@ namespace Amazon.PowerShell.Cmdlets.EMC
         {
             public System.String Description { get; set; }
             public System.String Name { get; set; }
+            public Dictionary<System.String, System.String> Tags { get; set; }
         }
         
     }