+++ modules/AWSPowerShell/Cmdlets/AmazonServerlessApplicationRepository/Basic/Get-SARApplication-Cmdlet.cs
@@ -42,7 +42,7 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter ApplicationId
         /// <summary>
         /// <para>
-        /// <para>The ID of the application to get.</para>
+        /// <para>The Amazon Resource Name (ARN) of the application.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]