+++ modules/AWSPowerShell/Cmdlets/AmazonServerlessApplicationRepository/Basic/Get-SARApplicationVersionList-Cmdlet.cs
@@ -44,7 +44,7 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter ApplicationId
         /// <summary>
         /// <para>
-        /// <para>The ID of the application to get.</para>
+        /// <para>The Amazon Resource Name (ARN) of the application.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]