+++ modules/AWSPowerShell/Cmdlets/AmazonServerlessApplicationRepository/Basic/Update-SARApplication-Cmdlet.cs
@@ -42,7 +42,7 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter ApplicationId
         /// <summary>
         /// <para>
-        /// <para>The ID of the application to get.</para>
+        /// <para>The Amazon Resource Name (ARN) of the application.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 0, ValueFromPipeline = true)]
@@ -52,7 +52,7 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter Author
         /// <summary>
         /// <para>
-        /// <para>The name of the author publishing the app.</para><para>Min Length=1. Max Length=127.</para><para>Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";</para>
+        /// <para>The name of the author publishing the app.</para><para>Minimum length=1. Maximum length=127.</para><para>Pattern "^[a-z0-9](([a-z0-9]|-(?!-))*[a-z0-9])?$";</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]
@@ -62,7 +62,7 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter Description
         /// <summary>
         /// <para>
-        /// <para>The description of the application.</para><para>Min Length=1. Max Length=256</para>
+        /// <para>The description of the application.</para><para>Minimum length=1. Maximum length=256</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]
@@ -83,7 +83,7 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter Label
         /// <summary>
         /// <para>
-        /// <para>Labels to improve discovery of apps in search results.</para><para>Min Length=1. Max Length=127. Maximum number of labels: 10</para><para>Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";</para>
+        /// <para>Labels to improve discovery of apps in search results.</para><para>Minimum length=1. Maximum length=127. Maximum number of labels: 10</para><para>Pattern: "^[a-zA-Z0-9+\\-_:\\/@]+$";</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]
@@ -94,8 +94,8 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter ReadmeBody
         /// <summary>
         /// <para>
-        /// <para>A raw text Readme file that contains a more detailed description of the application
-        /// and how it works in markdown language.</para><para>Max size 5 MB</para>
+        /// <para>A text readme file in Markdown language that contains a more detailed description
+        /// of the application and how it works.</para><para>Maximum size 5 MB</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]
@@ -105,8 +105,8 @@ namespace Amazon.PowerShell.Cmdlets.SAR
         #region Parameter ReadmeUrl
         /// <summary>
         /// <para>
-        /// <para>A link to the Readme file that contains a more detailed description of the application
-        /// and how it works in markdown language.</para><para>Max size 5 MB</para>
+        /// <para>A link to the readme file in Markdown language that contains a more detailed description
+        /// of the application and how it works.</para><para>Maximum size 5 MB</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter]